% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{acescg_conversion_matrices}
\alias{acescg_conversion_matrices}
\title{ACEScg conversion matrices}
\format{
Length-2 list of 3x3 matrices
}
\usage{
acescg_conversion_matrices
}
\description{
ACEScg conversion matrices
}
\keyword{datasets}
