% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_padding.R
\name{add_multi_padding}
\alias{add_multi_padding}
\title{add_multi_padding}
\usage{
add_multi_padding(input, pad = 1)
}
\arguments{
\item{input}{A two-dimensional matrix.}

\item{pad}{Number of padding entries}
}
\value{
Matrix with edges padded
}
\description{
Adds multiple levels padding to the matrix
}
\keyword{internal}
