% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_padding.R
\name{add_padding}
\alias{add_padding}
\title{add_padding}
\usage{
add_padding(input)
}
\arguments{
\item{input}{A two-dimensional matrix.}
}
\value{
Matrix with edges padded
}
\description{
Adds padding to the matrix
}
\keyword{internal}
