% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace_helpers.R
\name{apply_color_matrix}
\alias{apply_color_matrix}
\title{Apply a 3x3 color transform to an image}
\usage{
apply_color_matrix(img, M = diag(3))
}
\arguments{
\item{img}{3/4-channel array (RGB/A). 2D (grey) or 2-channel (grey+alpha) pass through.}

\item{M}{3x3 matrix.}
}
\value{
Array with transformed RGB; alpha preserved.
}
\description{
Apply a 3x3 color transform to an image
}
\keyword{internal}
