% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace_helpers.R
\name{array_rgb_to_hex_colmajor}
\alias{array_rgb_to_hex_colmajor}
\title{Array hex vector in column-major order}
\usage{
array_rgb_to_hex_colmajor(arr)
}
\arguments{
\item{arr}{Array with c>=3}
}
\value{
Hex vector length h*w
}
\description{
Array hex vector in column-major order
}
\keyword{internal}
