% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_bokeh_size.R
\name{calc_bokeh_size}
\alias{calc_bokeh_size}
\title{Calculate Bokeh Size}
\usage{
calc_bokeh_size(z, zfocus, f, N)
}
\arguments{
\item{z}{Depth matrix.}

\item{zfocus}{Depth in which to blur. Minimum 0, maximum 1.}

\item{f}{Focal length of the virtual camera. .}

\item{N}{F-stop. Focal length of the virtual camera.}
}
\value{
Matrix of bokeh sizes.
}
\description{
Calculates the amount of blurring at each point with camera characteristics.
}
\keyword{internal}
