% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{colorspace_to_exr_chromaticities}
\alias{colorspace_to_exr_chromaticities}
\title{Convert a colorspace descriptor to EXR chromaticities metadata}
\usage{
colorspace_to_exr_chromaticities(colorspace)
}
\arguments{
\item{colorspace}{Colorspace descriptor.}
}
\value{
Named EXR chromaticities list.
}
\description{
Convert a colorspace descriptor to EXR chromaticities metadata
}
\keyword{internal}
