% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace_helpers.R
\name{compute_cat_bradford}
\alias{compute_cat_bradford}
\title{Bradford CAT matrix from source to target white}
\usage{
compute_cat_bradford(
  source_xyz = c(0.95264, 1, 1.00827),
  target_xyz = c(0.95047, 1, 1.08883),
  eps = 1e-08
)
}
\arguments{
\item{source_xyz}{Default \code{c(0.95264,1,1.00827)}. Source XYZ white, Y=1.}

\item{target_xyz}{Default \code{c(0.95047,1,1.08883)}. Target XYZ white, Y=1.}

\item{eps}{Default \code{1e-8}. Small epsilon.}
}
\value{
3x3 numeric CAT matrix to apply in XYZ.
}
\description{
Bradford CAT matrix from source to target white
}
\keyword{internal}
