% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_color.R
\name{convert_color}
\alias{convert_color}
\title{Convert Color}
\usage{
convert_color(color, as_hex = FALSE, single = FALSE)
}
\arguments{
\item{color}{The color to convert. Can be either an R color string, or a
numeric scalar, RGB vector, RGBA vector, or RGB/RGBA matrix with values
between \code{0} and \code{1}.}

\item{as_hex}{Default \code{FALSE}. If \code{TRUE}, return hex colors.}

\item{single}{Default \code{FALSE}. If \code{TRUE}, require exactly one color.}
}
\value{
RGB or RGBA color vector, or a color matrix for multiple colors.
}
\description{
Convert Color
}
\examples{
#none
}
\keyword{internal}
