% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_native_raster.R
\name{convert_to_native_raster}
\alias{convert_to_native_raster}
\title{Convert to Native Raster}
\usage{
convert_to_native_raster(input)
}
\arguments{
\item{input}{Matrix or image array to convert to a native raster.}
}
\description{
Converts a matrix or image array to a native raster.
}
\keyword{internal}
