% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dragon}
\alias{dragon}
\title{Dragon Image}
\format{
An RGB 3-layer HDR array with 200 rows and 200 columns,
generated using the rayrender package.
}
\usage{
dragon
}
\description{
Dragon Image
}
\keyword{datasets}
