% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{exr_metadata_from_rayimg}
\alias{exr_metadata_from_rayimg}
\title{Build EXR metadata from a rayimg}
\usage{
exr_metadata_from_rayimg(image, metadata = NULL)
}
\arguments{
\item{image}{A \code{rayimg}.}

\item{metadata}{Default \code{NULL}. User-supplied EXR metadata.}
}
\value{
EXR metadata list.
}
\description{
Build EXR metadata from a rayimg
}
\keyword{internal}
