% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace_helpers.R
\name{get_whitepoint_xyz}
\alias{get_whitepoint_xyz}
\title{Normalize/label a white point}
\usage{
get_whitepoint_xyz(white_point = NULL)
}
\arguments{
\item{white_point}{Default \code{NULL}. Named ("D65","D60","D55","D50","D75","E") or XYZ (length 3, any Y).}
}
\value{
list(value = XYZ with Y=1, white_name = name or "Custom")
}
\description{
Normalize/label a white point
}
\keyword{internal}
