% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{libopenexr_supports_metadata}
\alias{libopenexr_supports_metadata}
\title{Check whether libopenexr supports metadata}
\usage{
libopenexr_supports_metadata()
}
\value{
\code{TRUE} if \code{libopenexr::write_exr()} accepts metadata.
}
\description{
Check whether libopenexr supports metadata
}
\keyword{internal}
