% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayimg.R
\name{make_colorspace}
\alias{make_colorspace}
\title{Build a colorspace descriptor}
\usage{
make_colorspace(
  name = "ACEScg",
  primaries = list(r = c(0.713, 0.293), g = c(0.165, 0.83), b = c(0.128, 0.044)),
  white_xy = c(0.32168, 0.33767),
  white_name = "D60"
)
}
\arguments{
\item{name}{Default \code{"ACEScg"}.}

\item{primaries}{Default AP1.}

\item{white_xy}{Default D60.}
}
\value{
list(name, rgb_to_xyz, xyz_to_rgb, white_xyz, white_name)
}
\description{
Build a colorspace descriptor
}
\keyword{internal}
