% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayimg.R
\name{make_rgb_xyz_matrices}
\alias{make_rgb_xyz_matrices}
\title{Build RGB-to-XYZ matrices from primaries + white}
\usage{
make_rgb_xyz_matrices(
  primaries = list(r = c(0.713, 0.293), g = c(0.165, 0.83), b = c(0.128, 0.044)),
  white_xy = c(0.32168, 0.33767)
)
}
\arguments{
\item{primaries}{Default \code{list(r=c(0.713,0.293), g=c(0.165,0.830), b=c(0.128,0.044))}. xy chromaticities (AP1).}

\item{white_xy}{Default \code{c(0.32168,0.33767)}. White xy (D60).}
}
\value{
List with 3x3 numeric matrices: \code{rgb_to_xyz}, \code{xyz_to_rgb}.
}
\description{
Build RGB-to-XYZ matrices from primaries + white
}
\keyword{internal}
