% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayimg.R
\name{new_rayimg}
\alias{new_rayimg}
\title{Create a rayimg}
\usage{
new_rayimg(x = NULL, filetype = NULL, gamma_correct = FALSE)
}
\arguments{
\item{x}{Default \code{NULL}. The underlying array/matrix to wrap.}

\item{filetype}{Default \code{NULL}. Original source type to record (e.g., "png").}

\item{gamma_correct}{Default \code{FALSE}. Whether to gamma correct the image.}
}
\value{
An object of class \verb{c("rayimg", <original class>)} with \code{filetype} attribute.
}
\description{
Create a rayimg
}
