% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{normalize_exr_chromaticities}
\alias{normalize_exr_chromaticities}
\title{Normalize EXR chromaticities metadata}
\usage{
normalize_exr_chromaticities(chromaticities)
}
\arguments{
\item{chromaticities}{EXR chromaticities metadata.}
}
\value{
Named chromaticities list.
}
\description{
Normalize EXR chromaticities metadata
}
\keyword{internal}
