% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_padding.R
\name{pad_to_fit}
\alias{pad_to_fit}
\title{pad to fit}
\usage{
pad_to_fit(dim, kernel)
}
\arguments{
\item{heightmap}{A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.}

\item{pad}{Number of padding entries}
}
\value{
Hillshade with edges trimmed
}
\description{
Pads to a rect
}
\keyword{internal}
