% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_asp_native_raster.R
\name{plot_asp_native_raster}
\alias{plot_asp_native_raster}
\title{Plot Native Raster with Custom Aspect Ratio}
\usage{
plot_asp_native_raster(
  nr,
  asp = 1,
  show = FALSE,
  return_grob = FALSE,
  gp = grid::gpar()
)
}
\description{
Plot Native Raster with Custom Aspect Ratio
}
\keyword{internal}
