% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image_grid.R
\name{plot_image_grid}
\alias{plot_image_grid}
\title{Plot Image Grid}
\usage{
plot_image_grid(input_list, dim = c(1, 1))
}
\arguments{
\item{input_list}{List of array (or matrix) image inputs.}

\item{...}{Additional arguments to pass to the \code{raster::plotRGB} function that displays the map.}
}
\description{
Displays the image in the current device.
}
\examples{
#if(interactive()){
#Plot the dragon array
plot_image_grid(list(dragon, dragon), dim = c(1,2))
plot_image_grid(list(dragon, dragon), dim = c(2,1))
plot_image_grid(list(dragon, dragon), dim = c(2,2))

#Plot alongside the depth matrix
dragon_depth_reoriented = render_reorient(dragondepth,
                                         transpose = TRUE,
                                         flipx = TRUE)/2000
plot_image_grid(list(dragondepth/2000, dragon, dragon, dragondepth/2000),
               dim = c(2,2))
#end}
}
