% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preview_focus.R
\name{preview_focus}
\alias{preview_focus}
\title{Preview Focus}
\usage{
preview_focus(image, depthmap, focus, imagetype, depthtype)
}
\arguments{
\item{image}{Image filename or 3-layer RGB array.}

\item{depthmap}{Depth map filename or 1d array.}

\item{focus}{Defaults \code{0.5}. Depth in which to blur. Minimum 0, maximum 1.}
}
\description{
Displays the focal point
}
\examples{
#Plot the dragon
}
\keyword{internal}
