% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{print.rayimg}
\alias{print.rayimg}
\title{Print method for rayimg}
\usage{
\method{print}{rayimg}(x = NULL, ..., decimals = 2)
}
\arguments{
\item{x}{Default \code{NULL}. A \code{rayimg} object.}

\item{...}{Default ``. Passed to downstream print methods.}

\item{decimals}{Default \code{2}. Number of decimal places to display.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for rayimg
}
