% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_helpers.R
\name{rayimg_composite_checker}
\alias{rayimg_composite_checker}
\title{Compose RGB over checkerboard using alpha}
\usage{
rayimg_composite_checker(rgb, alpha = NULL, light = 0.92, dark = 0.75)
}
\description{
Compose RGB over checkerboard using alpha
}
\keyword{internal}
