% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayimg.R
\name{rayimg_mark_grey}
\alias{rayimg_mark_grey}
\title{(internal) Mark rayimg as grayscale}
\usage{
rayimg_mark_grey(x)
}
\description{
Tag an image so the print method treats it as a single Grey channel,
even if the underlying data has RGB duplicates and/or alpha.
}
\keyword{internal}
