% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_helpers.R
\name{rayimg_vertical_ellipsis}
\alias{rayimg_vertical_ellipsis}
\title{Vertical ellipsis symbol}
\usage{
rayimg_vertical_ellipsis(use_cli = TRUE)
}
\arguments{
\item{use_cli}{Default \code{TRUE}. If \code{TRUE}, consult \code{cli::is_utf8_output()} for Unicode support.}
}
\value{
"⋮" if Unicode likely supported, otherwise "...".
}
\description{
Vertical ellipsis symbol
}
\keyword{internal}
