% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_bokeh.R
\name{render_bokeh}
\alias{render_bokeh}
\title{render_bokeh}
\usage{
render_bokeh(image, depthmap, focus = 0.5, focallength = 1,
  fstop = 1, filename = NULL, bokehshape = "circle",
  bokehintensity = 5, rotation = 0, progress = TRUE)
}
\arguments{
\item{image}{Image filename.}

\item{depthmap}{Depth map filename.}

\item{focus}{Defaults `0.5`. Depth in which to blur. Minimum 0, maximum 1.}

\item{focallength}{Default `1`. Focal length of the virtual camera.}

\item{fstop}{Default `1`. F-stop of the virtual camera.}

\item{filename}{The filename of the image to be saved. If this is not given, the image will be plotted instead.}

\item{bokehshape}{Default `circle`. Also built-in: `hex`. The shape of the bokeh.}

\item{bokehintensity}{Default `3`. Intensity of the bokeh when the pixel intensity is greater than `bokehlimit`.}

\item{rotation}{Default `0`. Number of degrees to rotate the hexagon bokeh shape.}

\item{bokehlimit}{Default `0.8`. Limit after which the bokeh intensity is increased by `bokehintensity`.}

\item{gamma_correction}{Default `TRUE`. Controls gamma correction when adding colors. Default exponent of 2.2.}
}
\description{
Takes an image and a depth map to render the image with depth of field.
}
\examples{
filename_map = tempfile()
}
