% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_to_display.R
\name{render_to_display}
\alias{render_to_display}
\title{Render To Display}
\usage{
render_to_display(image, display = CS_SRGB, encode = TRUE, adapt_white = TRUE)
}
\arguments{
\item{image}{A \code{rayimg}, array, or filename.}

\item{display}{Default \code{CS_SRGB}. Target display RGB space.}

\item{encode}{Default \code{TRUE}. If \code{TRUE}, apply sRGB OETF (only valid when \code{display} is sRGB-like).}

\item{adapt_white}{Default \code{TRUE}. Whether to perform CAT from working white to display white.}
}
\value{
A \code{rayimg} when \code{encode=FALSE}, or a \code{rayimg} with sRGB-encoded RGB when \code{encode=TRUE}.
}
\description{
Convert an image from its working space to a display space, and optionally apply sRGB OETF for viewing.
}
