% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorspace_helpers.R
\name{rgb_to_hex}
\alias{rgb_to_hex}
\title{Encode RGB/RGBA to hex (sRGB domain)}
\usage{
rgb_to_hex(x)
}
\arguments{
\item{x}{Values. 3 or 4 columns (R,G,B,A) in 0 to 1.}
}
\value{
Hex vector (#RRGGBB or #RRGGBBAA).
}
\description{
Encode RGB/RGBA to hex (sRGB domain)
}
\keyword{internal}
