% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_image_array.R
\name{rotate_image_array}
\alias{rotate_image_array}
\title{Rotate an image array around its center}
\usage{
rotate_image_array(img, angle)
}
\description{
Rotate an image array around its center
}
\keyword{internal}
