% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{srgb_conversion_matrices}
\alias{srgb_conversion_matrices}
\title{sRGB conversion matrices}
\format{
Length-2 list of 3x3 matrices
}
\usage{
srgb_conversion_matrices
}
\description{
sRGB conversion matrices
}
\keyword{datasets}
