% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{[.rayimg}
\alias{[.rayimg}
\title{Subset a rayimg without losing attributes}
\usage{
\method{[}{rayimg}(x = NULL, ..., drop = TRUE)
}
\arguments{
\item{x}{Default \code{NULL}. A \code{rayimg} object.}

\item{...}{Default ``. Standard subsetting indices.}

\item{drop}{Default \code{TRUE}. Whether to drop dimensions (as in base \code{[}).}
}
\value{
A subset of \code{x}. Preserves the \code{filetype} attribute. Preserves the
\code{rayimg} class if the result still has dimensions (array/matrix).
}
\description{
Subset a rayimg without losing attributes
}
