% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sunset_image}
\alias{sunset_image}
\title{Sunset Image}
\format{
An RGBA 4-layer LDR rayimg array from a JPEG image
with 400 rows and 400 columns.
}
\usage{
sunset_image
}
\description{
Sunset Image
}
\author{
Tyler Morgan-Wall
}
\keyword{datasets}
