% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_padding.R
\name{trim_padding}
\alias{trim_padding}
\title{trim_padding}
\usage{
trim_padding(input, pad = 1)
}
\arguments{
\item{input}{A two-dimensional matrix.}

\item{pad}{Number of padding entries}
}
\value{
Matrix with edges trimmed
}
\description{
Trims padding
}
\keyword{internal}
