% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{white_current_from_exr_metadata}
\alias{white_current_from_exr_metadata}
\title{Get the adopted neutral white point from EXR metadata}
\usage{
white_current_from_exr_metadata(metadata)
}
\arguments{
\item{metadata}{EXR metadata list.}
}
\value{
XYZ white point, or \code{NULL} when absent.
}
\description{
Get the adopted neutral white point from EXR metadata
}
\keyword{internal}
