% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{xy_to_xyz}
\alias{xy_to_xyz}
\title{Convert xy chromaticity coordinates to XYZ}
\usage{
xy_to_xyz(xy)
}
\arguments{
\item{xy}{xy chromaticity vector.}
}
\value{
XYZ vector normalized to Y=1.
}
\description{
Convert xy chromaticity coordinates to XYZ
}
\keyword{internal}
