% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exr_metadata.R
\name{xyz_to_xy}
\alias{xyz_to_xy}
\title{Convert XYZ to xy chromaticity coordinates}
\usage{
xyz_to_xy(xyz)
}
\arguments{
\item{xyz}{XYZ vector.}
}
\value{
Numeric xy vector.
}
\description{
Convert XYZ to xy chromaticity coordinates
}
\keyword{internal}
