% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_title.R
\name{add_title}
\alias{add_title}
\title{Add Title Function (deprecated)}
\usage{
add_title(...)
}
\arguments{
\item{...}{Arguments to pass to \code{render_title()} function.}
}
\value{
A 3-layer RGB array of the processed image if \code{filename = NULL} and \code{preview = FALSE}.
Otherwise, writes the image to the specified file or displays it if \code{preview = TRUE}.
}
\description{
Add Title Function (deprecated)
}
\examples{
#Deprecated add_title function. Will display a warning.
if(run_documentation()) {
render_title(dragon, preview = TRUE, title_text = "Dragon", title_size=20)
}
}
