% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_vignette.R
\name{add_vignette}
\alias{add_vignette}
\title{Add Vignette Effect (Deprecated)}
\usage{
add_vignette(...)
}
\arguments{
\item{...}{Arguments to pass to \code{render_title()} function.}
}
\value{
3-layer RGB array of the processed image.
}
\description{
Takes an RGB array/filename and adds a camera vignette effect.
}
\examples{
if(run_documentation()){
#Plot the dragon
plot_image(dragon)
}
if(run_documentation()){
#Add a vignette effect:
add_vignette(dragon, preview = TRUE, vignette = 0.5)
}
}
