% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_native_raster.R
\name{convert_to_native_raster}
\alias{convert_to_native_raster}
\title{Plot Image}
\usage{
convert_to_native_raster(input)
}
\arguments{
\item{input}{Array to get converted to a native raster}

\item{ncol}{Number of columns}

\item{nrow}{Number of rows}
}
\description{
Displays the image in the current device.
}
\keyword{internal}
