% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_2d_exponential.R
\name{generate_2d_exponential}
\alias{generate_2d_exponential}
\title{Generate 2D exponential Distribution}
\usage{
generate_2d_exponential(falloff = 1, dim = 11, width = 3)
}
\arguments{
\item{falloff}{Default \code{1}. Falloff of the exponential.}

\item{dim}{Default \code{11}. The dimensions of the resulting square matrix.}

\item{width}{Default \code{3} (\code{-10} to \code{10}). The range in which to compute the distribution.}
}
\description{
Generates a 2D exponential distribution, with an optional
argument to take the exponential to a user-defined power.
}
\examples{
image(generate_2d_exponential(1,31,3), asp=1)
}
