% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_image.R
\name{plot_image}
\alias{plot_image}
\title{Plot Image}
\usage{
plot_image(
  input,
  rotate = 0,
  keep_user_par = FALSE,
  asp = 1,
  new_page = TRUE,
  return_grob = FALSE,
  ...
)
}
\arguments{
\item{input}{Image or filename of an image to be plotted.}

\item{rotate}{Default 0. Rotates the output. Possible values: 0, 90, 180, 270.}

\item{keep_user_par}{Default \code{TRUE}. Whether to keep the user's \code{par()} settings. Set to \code{FALSE} if you
want to set up a multi-pane plot (e.g. set \code{par(mfrow)}).}

\item{asp}{Default \code{1}. Aspect ratio of the pixels in the plot. For example, an aspect ratio of \code{4/3} will
slightly widen the image.}

\item{new_page}{Default \code{TRUE}. Whether to call \code{grid::grid.newpage()} before plotting the image.}

\item{return_grob}{Default \code{FALSE}. Whether to return the grob object.}

\item{...}{Additional arguments to pass to the \code{raster::plotRGB} function that displays the map.}
}
\description{
Displays the image in the current device.
}
\examples{
#if(interactive()){
#Plot the dragon array
plot_image(dragon)
#end}
}
