% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rayimg}
\alias{print.rayimg}
\title{Print method for rayimg}
\usage{
\method{print}{rayimg}(x = NULL, preview_n = 10, decimals = 3, color = TRUE, ...)
}
\arguments{
\item{x}{Default \code{NULL}. A \code{rayimg} object.}

\item{preview_n}{Default \code{10}. Max rows/cols to display in numeric preview.}

\item{decimals}{Default \code{3}. Number of decimal places to display in numeric preview.}

\item{color}{Default \code{TRUE}. Colorize headers and R/G/B/A numeric cells when supported.#' @param ... Default ``. Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for rayimg
}
