% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rayimg.R
\name{rayimg}
\alias{rayimg}
\title{Create a rayimg}
\usage{
rayimg(x = NULL, filetype = NULL, gamma_correct = FALSE)
}
\arguments{
\item{x}{Default \code{NULL}. The underlying array/matrix to wrap.}

\item{filetype}{Default \code{NULL}. Original source type to record (e.g., "png").}

\item{gamma_correct}{Default \code{FALSE}. Whether to gamma correct the image.}
}
\value{
An object of class \verb{c("rayimg", <original class>)} with \code{filetype} attribute.
}
\description{
This class will always be:
\itemize{
\item A matrix or a 2/3/4 layer array
\itemize{
\item 1 Layer(s): c("Greyscale")
\item 2 Layer(s): c("Greyscale", "Alpha")
\item 3 Layer(s): c("Red", "Green", "Blue")
\item 4 Layer(s): c("Red", "Green", "Blue", "Alpha")
}
\item It tracks whether the image needs to be gamma corrected for display
\item 
}
}
\keyword{internal}
