% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_white_balance.R
\name{render_white_balance}
\alias{render_white_balance}
\title{Render White Balance (Bradford CAT)}
\usage{
render_white_balance(
  image,
  reference_white = NA,
  target_white = "D60",
  filename = NULL,
  preview = FALSE
)
}
\arguments{
\item{image}{Default \code{NULL}. 3-layer RGB/4-layer RGBA array, \code{rayimg}, or filename.}

\item{reference_white}{Default \code{NA}, automatically pulled from image if available. Source white (Y=1 XYZ or named: "D60","D65","D50","E"...).}

\item{target_white}{Default \code{"D60"}. Target white (same format as \code{reference_white}).}

\item{filename}{Default \code{NULL}. Output path.}

\item{preview}{Default \code{FALSE}. If \code{TRUE}, display the image.}
}
\value{
A \code{rayimg} RGBA array.
}
\description{
Chromatic-adapt an image from \code{reference_white} to \code{target_white} in the working space.
}
\examples{
# Not specifying a target white balance returns an unchanged image
render_white_balance(dragon, preview = TRUE)

# Default white "D60", converting to D50 produces a warmer image
render_white_balance(dragon, preview = TRUE, target_white = "D50")

# Default white "D60", converting to "D75" produces a cooler image
render_white_balance(dragon, preview = TRUE, target_white = "D75")

# Set reference white to colder "D75", converting to "D50" produces a very warm image
render_white_balance(dragon, preview = TRUE,
                     reference_white= "D75", target_white = "D50")

# With a real, non-rendered image:
render_white_balance(sunset_image, preview=TRUE)

# Cooler
render_white_balance(sunset_image, preview=TRUE, target_white = "D75")

# Warmer
render_white_balance(sunset_image, preview=TRUE, target_white = "D50")
}
